theme: Move theme handling to own file.
