+++ pywal/export.py
@@ -53,7 +53,6 @@ def every(colors, output_dir=CACHE_DIR):
     colors = flatten_colors(colors)
     template_dir = os.path.join(MODULE_DIR, "templates")
     template_dir_user = os.path.join(CONF_DIR, "templates")
-    util.create_dir(template_dir_user)
 
     join = os.path.join  # Minor optimization.
